@@ -40,45 +34,34 @@
     "mdbreact": "^5.0.1",
     "moment": "^2.24.0",
     "rc-virtual-list": "^3.10.5",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.0.0-beta.16",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.2",
     "react-bootstrap-icons": "^1.3.0",
-    "react-color": "^2.19.3",
-    "react-countdown": "^2.3.2",
     "react-countdown-circle-timer": "^3.2.1",
-    "react-datepicker": "^4.6.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-drag-drop-files": "^2.3.7",
-    "react-easy-crop": "^4.7.4",
-    "react-icons": "^4.4.0",
+    "react-easy-crop": "^5.0.6",
+    "react-helmet": "^6.1.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-json-view": "^1.21.3",
-    "react-multi-email": "^0.5.3",
-    "react-redux": "^7.1.3",
     "react-router-dom": "^5.1.2",
     "react-rte": "^0.16.5",
-    "react-scripts": "3.4.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.3.2",
     "react-select-country-list": "^2.2.3",
-    "react-slick": "^0.29.0",
-    "react-sortable-tree-patch-react-17": "2.9.0",
-    "react-spring": "^9.7.2",
-    "react-toastify": "^8.0.2",
-    "react-tooltip": "^4.2.21",
+    "react-slick": "^0.30.2",
     "react-transition-group": "^4.4.5",
     "react-type-animation": "^3.0.1",
-    "react-window": "^1.8.8",
     "recharts": "^2.7.3",
-    "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0",
-    "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
-    "tooltip": "^1.6.1",
     "validator": "^13.7.0"
   },
   "scripts": {
-    "start": "env-cmd -f .env.production react-scripts --max-old-space-size=6144 start",
-    "build": "env-cmd -f .env.production react-scripts --max-old-space-size=6144 build",
+    "start": "cross-env NODE_OPTIONS=--max-old-space-size=6144 react-scripts start",
+    "build": "cross-env NODE_OPTIONS=--max-old-space-size=6144 react-scripts build",
+    "start:vistaar": "cross-env REACT_APP_INSTANCE=VISTAAR REACT_APP_BASEURL=https://dev-vistaar.digitalgreen.org/be/ REACT_APP_BASEURL_without_slash=https://dev-vistaar.digitalgreen.org/be NODE_OPTIONS=--max-old-space-size=6144 react-scripts start",
+    "start:eadp": "cross-env REACT_APP_INSTANCE=EADP REACT_APP_BASEURL=https://datahubethdev.farmstack.co/be/ REACT_APP_BASEURL_without_slash=https://datahubethdev.farmstack.co/be NODE_OPTIONS=--max-old-space-size=6144 react-scripts start",
+    "start:kadp": "cross-env REACT_APP_INSTANCE=KADP REACT_APP_BASEURL=https://datahubethdev.farmstack.co/be/ REACT_APP_BASEURL_without_slash=https://datahubethdev.farmstack.co/be NODE_OPTIONS=--max-old-space-size=6144 react-scripts start",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "coverage": "react-scripts test --coverage --watchAll"
